requirements.txt
Flask==0.10.1
Flask-Markdown==0.3
Flask-SQLAlchemy==1.0
Jinja2==2.7.1
Markdown==2.3.1
MarkupSafe==0.18
SQLAlchemy==0.8.2
Werkzeug==0.9.3
argparse==1.2.1
blinker==1.3
distribute==0.6.34
gunicorn==17.5
itsdangerous==0.23
psycopg2==2.5.1
python-postmark==0.3.2
wsgiref==0.1.2
